requirements.txt
sam2
torch
torchvision
numpy
Pillow
matplotlib
